+++ backend/src/test/java/codesquard/app/api/item/ItemServiceTest.java
@@ -1,5 +1,6 @@
 package codesquard.app.api.item;
 
+import static codesquard.app.CategoryTestSupport.*;
 import static codesquard.app.ImageTestSupport.*;
 import static codesquard.app.MemberTestSupport.*;
 import static codesquard.app.RegionTestSupport.*;
@@ -17,6 +18,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import org.assertj.core.api.Assertions;
+import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -29,12 +32,13 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import codesquard.app.CategoryTestSupport;
+import codesquard.app.api.errors.exception.RestApiException;
 import codesquard.app.api.image.ImageUploader;
 import codesquard.app.api.item.request.ItemModifyRequest;
 import codesquard.app.api.item.request.ItemRegisterRequest;
-import codesquard.app.api.response.ItemResponse;
-import codesquard.app.api.response.ItemResponses;
+import codesquard.app.api.item.response.ItemDetailResponse;
+import codesquard.app.api.item.response.ItemResponse;
+import codesquard.app.api.item.response.ItemResponses;
 import codesquard.app.domain.category.Category;
 import codesquard.app.domain.category.CategoryRepository;
 import codesquard.app.domain.image.Image;
@@ -49,6 +53,8 @@ import codesquard.app.domain.membertown.MemberTownRepository;
 import codesquard.app.domain.oauth.support.Principal;
 import codesquard.app.domain.region.Region;
 import codesquard.app.domain.region.RegionRepository;
+import codesquard.app.domain.wish.Wish;
+import codesquard.app.domain.wish.WishRepository;
 import codesquard.support.SupportRepository;
 
 @SpringBootTest
@@ -78,6 +84,9 @@ class ItemServiceTest {
 	@Autowired
 	private ImageRepository imageRepository;
 
+	@Autowired
+	private WishRepository wishRepository;
+
 	@MockBean
 	private ImageUploader imageUploader;
 
@@ -86,6 +95,7 @@ class ItemServiceTest {
 
 	@AfterEach
 	void tearDown() {
+		wishRepository.deleteAllInBatch();
 		imageRepository.deleteAllInBatch();
 		itemRepository.deleteAllInBatch();
 		categoryRepository.deleteAllInBatch();
@@ -158,7 +168,7 @@ class ItemServiceTest {
 	@Test
 	public void modifyItem() throws IOException {
 		// given
-		Category category = categoryRepository.save(CategoryTestSupport.findByName("스포츠/레저"));
+		Category category = categoryRepository.save(findByName("스포츠/레저"));
 		Member member = memberRepository.save(createMember("avatarUrlValue", "23Yong@gmail.com", "23Yong"));
 
 		Region region = regionRepository.save(createRegion("서울 송파구 가락동"));
@@ -231,5 +241,66 @@ class ItemServiceTest {
 
 		// then
 		assertThat(updateItem.getStatus()).isEqualTo(ItemStatus.RESERVED);
+	}
+
+	@DisplayName("판매자가 한 상품의 상세한 정보를 조회한다")
+	@Test
+	public void findDetailItemBySeller() {
+		// given
+		Member member = memberRepository.save(createMember("avatarUrlValue", "23Yong@gmail.com", "23Yong"));
+		categoryRepository.saveAll(getCategories());
+		Category category = findByName("스포츠/레저");
+		Item item = Item.builder()
+			.title("빈티지 롤러 블레이드")
+			.content("어린시절 추억의향수를 불러 일으키는 롤러 스케이트입니다.")
+			.price(200000L)
+			.status(ON_SALE)
+			.region("가락동")
+			.createdAt(now())
+			.wishCount(0L)
+			.viewCount(0L)
+			.chatCount(0L)
+			.member(member)
+			.category(category)
+			.build();
+		Item saveItem = itemRepository.save(item);
+
+		List<Image> images = List.of(
+			new Image("imageUrlValue1", new Item(saveItem.getId())),
+			new Image("imageUrlValue2", new Item(saveItem.getId())));
+		imageRepository.saveAll(images);
+
+		Wish wish = new Wish(member, item, now());
+		wishRepository.save(wish);
+
+		// when
+		ItemDetailResponse response = itemService.findDetailItemBy(item.getId(), member.getId());
+		// then
+		SoftAssertions.assertSoftly(softAssertions -> {
+			softAssertions.assertThat(response)
+				.extracting("isSeller", "seller", "status", "title", "categoryName", "content", "chatCount",
+					"wishCount", "viewCount", "price")
+				.contains(true, "23Yong", "판매중", "빈티지 롤러 블레이드", "스포츠/레저", "어린시절 추억의향수를 불러 일으키는 롤러 스케이트입니다.", 0L, 0L, 0L,
+					200000L);
+			softAssertions.assertThat(response.getImageUrls())
+				.hasSize(2);
+			softAssertions.assertAll();
+		});
+	}
+
+	@DisplayName("존재하지 않는 상품 등록번호로 상품을 조회할 수 없다")
+	@Test
+	public void findDetailItemWithNotExistItem() {
+		// given
+		Member member = createMember("avatarUrlValue", "23Yong@gmail.com", "23Yong");
+		Long itemId = 9999L;
+		// when
+		Throwable throwable = Assertions.catchThrowable(
+			() -> itemService.findDetailItemBy(itemId, member.getId()));
+		// then
+		Assertions.assertThat(throwable)
+			.isInstanceOf(RestApiException.class)
+			.extracting("errorCode.message")
+			.isEqualTo("상품을 찾을 수 없습니다.");
 	}
 }
